move: 상담 관련 파일 정리 작업
