+++ daehan_univ/src/main/java/kr/ac/dhuniv/counsel/controller/AdminCounselController.java
@@ -1,4 +1,4 @@
-package kr.ac.dhuniv.counsel.controller.admin;
+package kr.ac.dhuniv.counsel.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
